==> 03_code/02_01_Modelling31Aug2014.Rmd <==
---
output:
  pdf_document: default
  html_document: default
---
This file imports the means of the streams of the 5 decks from each of the two screens in the plant,
Then applies (maps) the models over each stream.
Most of PSD graphs are produced in this code. All charts are saved as immages for use in writeup (base.RMd)
An output is produced that is used in the following section "Model Selection"

import from 01_ImportWrangle2Means19Oct2014.Rmd

```{r, include = FALSE}
library("here")
source(here::here("02_rpackages",'projectpackages.R'))

knitr::opts_chunk$set(echo = FALSE, warning=FALSE,message=FALSE)
knitr::opts_chunk$set(fig.align="center") #default align for figures
knitr.table.format = "latex" #global option for tables. can still set any in each kable function.
```


import wrangled means
```{r}
unmod2 <- read_csv(here("03_code/01_code-output","RawPSD.csv"))
```


```{r}
# change underpan from 0 to 1 micron to avoid log0 
#im sure this step can be removed as this line will be remove for the moddeling anyway...
  tmp <- rev(unmod2$sieve)
 if(tmp[1]  < 1) {
    tmp[1] <- 1
    unmod2$sieve <- rev(tmp)
  }

```




```{r}
unmodlong2 <- unmod2 %>% 
  pivot_longer(-sieve, names_to= "stream" , values_to = "probability")%>% 
 group_by(stream) %>% 
  mutate (cumret = cumsum(probability))


#setting values that supposed t obe "1" to "1"

  tmp <- rev(unmodlong2$cumret)
 if(tmp[1] - 1 < 1e-12) {
    tmp[1] <- 1
    unmodlong2$cumret <- rev(tmp)
  }

unmodlong2$cumpass <- (1-unmodlong2$cumret)  


unmodlongDF <- data.frame(unmodlong2,
                             Df = rep(c("Feed (unmodeled)","OS (unmod.)","US (unmod.)" ),                                                                         times=c(2,2,2))) #the 222 :the df has 2 feeds 2 oversize 2 undersize factors that gets repeated.
```


```{r}
(plotUnmod <- ggplot(unmodlongDF, aes(sieve, probability, colour = stream)) +
  geom_line(size = 1) +
  geom_point() +
  ggtitle(" Measured (Unmoddeled) Probability PSDs for both Screens 13 and -14.") +
  theme(plot.title = element_text(size=9)) +
  facet_wrap(~Df))


(plotUnmodCum <- ggplot(unmodlongDF, aes(sieve, cumpass, colour = stream)) +
  geom_line(size = 1) +
  geom_point() +
  ggtitle(" Measured (Unmoddeled) Cummulative probabillity PSDs for both Screens 13 and -14.") +
  theme(plot.title = element_text(size=9)) +
  facet_wrap(~Df))
# Cummulative mass Passing = 
ggsave(here("05_fig_output","allStreamsPreModelling.pdf"), plotUnmod, width=10, height=8)

ggsave(here("05_fig_output","allStreamsCummulativePreModelling.pdf"), plotUnmodCum, width=10, height=8)
```


```{r}
#remove last row in wide tibble before converting to tidy datat with subsequint cummulative calc

cumsumshort <- unmod2 %>% 
  dplyr::select(-sieve) %>% 
  cumsum() %>% 
  dplyr::slice(1:11) 

sieveSliced<- unmod2 %>% 
  dplyr::select(sieve) %>% 
  dplyr::slice(1:11)

#join the x and y back into one tibble
SlicedCumsum <- cbind(sieveSliced ,cumsumshort)
```


```{r}
#Pivot longer then group_by stream and calculate the cumulative %retained "cumret".

longCumsum <- SlicedCumsum %>% 
  pivot_longer(-sieve, names_to= "stream" , values_to = "cumret")%>% 
 group_by(stream) %>% 
  mutate (cumpass = 1- cumret)

```


```{r}
#Rosin Rammler plot

longCumsumDF <- data.frame(
  longCumsum, Df = rep(c("Feed (modeled)","OS (mod.)","US (mod.)" ), 
                       times=c(2,2,2)))

write.csv(longCumsumDF, here::here("03_code/01_code-output","longCumsumDF.csv"))


(plotModRR <- ggplot(longCumsumDF, 
                        aes(log(sieve), log(-log(1-cumpass)), colour = stream)) +
  # geom_line(size = 1) +
  geom_point() +
  geom_smooth(method = "lm", se = FALSE)+
  ggtitle(" Rosin Rammler Modelled distributions.") +
  theme(plot.title = element_text(size=9)) +
  facet_wrap(~Df))


```


```{r}
longCumsumDF <- data.frame(
  longCumsum, Df = rep(c("Feed (modeled)","OS (mod.)","US (mod.)" ), times=c(2,2,2)))

(plotModGGS <- ggplot(longCumsumDF, 
                        aes(log(sieve), log(cumpass), colour = stream)) +
  # geom_line(size = 1) +
  geom_point() +
  geom_smooth(method = "lm", se = FALSE)+
  ggtitle(" Gates-Gaudin-Schumann Modelled distributions.") +
  theme(plot.title = element_text(size=9)) +
  facet_wrap(~Df))



ggsave(here("05_fig_output","RRModelled.pdf"), plotModRR, width=10, height=8)

ggsave(here("05_fig_output","GGSModelled.pdf"), plotModGGS, width=10, height=8)

```


```{r}
c <- 1e-12


#functions
RosinRammler <- function(df) {
  lm(log(-log(1-cumpass + c)) ~ log(sieve),
    data = df
    )
}

GatesGaudinSchumann <- function(df) {
  lm(log(cumpass) ~ log(sieve), 
     data = df
    )
}
```





trying to replace the model function above ..the "df" part is confusing where does the data come from.
so now below i try hadleys way :P



```{r}
Reg <- longCumsumDF %>% 
  group_by(stream) %>% 
  nest() %>% 
 mutate(
   RRmodels = lapply(data, RosinRammler),# "RRmodels" column will be created in the mutate. RosinRammler defined function is applied to the data "data" within (pipe opperator)  "longCumSumDF" the data frame.
   tidied_RR = map (RRmodels, tidy),
   glanced_RR = map(RRmodels, glance),
   augmented_RR = map (RRmodels, augment),

  #GGS model
   GGS_models = lapply(data, GatesGaudinSchumann),# "GGS_models" column will be created in the mutate.
   tidied_GGS = map (GGS_models, tidy),
   glanced_GGS = map(GGS_models, glance))#,
   # augmented_GGS = map (GGS_models, augment)) %>% 
   #  mutate_if(is.numeric, list(~na_if(., abs(Inf))))
```
```{r}
saveRDS(Reg, file = (here::here("03_code/01_code-output","Reg.Rdata"))) #use saveRDS as im saving a more complex data group or immage.
```








Model coefficients (slope and intercept) of fit

```{r}
tidiedRR <- Reg %>% 
  unnest(tidied_RR) %>% 
  dplyr::select(stream,term, estimate)

tidiedGGS <- Reg %>% 
  unnest(tidied_GGS) %>% 
  dplyr::select(stream,term, estimate)

write.csv(tidiedRR, file = here("03_code/01_code-output","CoefRR.csv"), row.names = FALSE)
write.csv(tidiedGGS, file = here("03_code/01_code-output","CoefGGS.csv"), row.names = FALSE)
```






statistical measures of model fit.
CLeaning up the glanced data and only selecting the output required for presentation.
Coefficient of determination `R^2`.(i tried latex formatting for the R2 not yet successful)







rsquared data kept here
```{r}
glancedRR <- Reg %>% 
  unnest(glanced_RR)

glancedGGS <- Reg %>% 
  unnest(glanced_GGS)
```


```{r}
#coeficient of determinant
R2RR <- glancedRR %>% 
      dplyr::select(stream, R2 = r.squared, adj.R2 = adj.r.squared, p.value, AIC)#conduct select and rename in one step with dplyr::select()

R2GGS <- glancedGGS %>% 
      dplyr::select(stream, R2 = r.squared, adj.R2 = adj.r.squared, p.value, AIC)

write.csv(R2RR, here("03_code/01_code-output","R2RR.csv"),row.names = FALSE)
write.csv(R2GGS, here("03_code/01_code-output","R2GGS.csv"),row.names = FALSE)
#written csvs are draw in by ~03_code/tableR2.Rmd

```



```{r}

```
